requirements.txt
speedtest-cli==2.1.3
ping3==2.7.0
python-dotenv==0.17.0
influxdb-client==1.16.0